BCDA-9515: Update handling BFD response errors (#1233)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9515

## 🛠 Changes

- Check for error returned from `c.Do` instead of nil resp, since resp
body is ignored on errors.
- Remove manual closing of resp body when error is not nil; per the HTTP
docs: `The request Body, if non-nil, will be closed by the underlying
Transport, even on errors. The Body may be closed asynchronously after
Do returns. On error, any Response can be ignored`
[Source](https://pkg.go.dev/net/http#Client.Do)

## ℹ️ Context

Unable to reach BFD v3 environments when testing and error is not being
surfaced when network connectivity problems occur.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Tests pass.