BCDA-9549: Update v3 test aco runout filename (#1243)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9549

## 🛠 Changes

Runout filename updated to runout convention.

## ℹ️ Context

The runout file for v3 was having its performance year updated from PY-1
to the current PY, due to the incorrect naming convention. Updated the
name to use the runout convention so it's not included in the
attribution refresh.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

smoke tests passing for
https://github.com/CMSgov/bcda-app/actions/runs/19045403895?pr=1224 when
PY is current performance year - 1.