BCDA-9363: reconfigure various volumes and directories in docker (#1245)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9363

## 🛠 Changes

- copy configs into docker container
- change directory of efs mount to match that of deployed envs
- change expected path of FHIR_TEMP_DIR to match that of deployed envs
- explicitly name docker volumes in dockerfiles
- remove legacy "ETL mode" configuration

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

Something to look at carefully: are the updates to `main.go` safe? For
context, the local environment runs in API mode (`BCDA_ETL_MODE` =
false) while all deployed envs run in ETL mode (`BCDA_ETL_MODE` = true).
I suspect that this won't be a problem, since the function that creates
the directory doesn't do anything if the directory doesn't already
exist.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->
Tested in dev