BCDA-9578: Refactor how cleanupjob gets slack token (#1249)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9578

## 🛠 Changes

Refactor how worker CleanupJob gets access to slack token.

## ℹ️ Context

Previously we were requiring AWS access to param store via local
context. This was affected by the AWS v2 migration work. Workers do not
have the proper context so this was failing.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.