BCDA-9570: fix ecr workflows and adjust logging (#1250)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9570

## 🛠 Changes

- set go version for golangci-lint pre-commit hook so that updates don't
break local development in an opaque way
- move slog logger for river to log package and refactor loggers a
little
- allow publish workflows to publish tags to non-prod ECR

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

- Devs intermittently encounter some issues with syncing the repo golang
version with that which builds golangci-lint for a pre-commit hook. The
update here specifies which golang version to build golangci-lint, which
should help avoid these issues.
- River requires slog, a different kind of logger from that of the rest
of the application. This PR moves the creation/config of the slog logger
to the log package and refactors functions there a little to make things
slightly more reusable/testable.
- Publish workflows are currently unable to push tags to the non-prod
account, and this PR enables addresses that.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

pre-commit changes tested locally
logging changes tested with unit tests