BCDA-9581: Add max attempts to river (#1253)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9581

## 🛠 Changes

Set river queue max attempts to 10. This is roughly a few hours of
attempts (as delayed by an exponential backoff mechanism).

## ℹ️ Context

Currently the max attempts is 25 (default). With the exponential backoff
this is up to 3 weeks of retries... That is far too much for all of our
jobs.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing.