BCDA-9612: Refactor smoke test logic checks (#1257)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9612

## 🛠 Changes

Refactor smoke test logic checks to make sure certain sections are
running as expected.

## ℹ️ Context

Recently we noticed that certain sections of our smoke tests werent
running as expected. We believe this is due to a very recent change in
refactoring if logic.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Smoke test workflow was run from PR branch and was running as expected.

---------

Co-authored-by: laurenkrugen-navapbc <126501259+laurenkrugen-navapbc@users.noreply.github.com>