BCDA-9594: update docker fhir temp dir path (#1262)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9594

## 🛠 Changes

- Instances of `TEMP_FHIR_DIR` have been updated to `FHIR_TEMP_DIR`

## ℹ️ Context

Env variable and actual directory path of `TEMP_FHIR_DIR` and
`FHIR_TEMP_DIR` have made for confusing changes; updating to
FHIR_TEMP_DIR to make both consistent and so that we can inject in
container setup. EC2 instances will be using the old env variable and we
should not update the env variable value; this change affects ECS only.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Verified changes working in dev - able to start export job and job
processes successfully.