BCDA-9617: Add region explicitly to request (#1266)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9617

## 🛠 Changes

Added explicit region to actual AWS request.

## ℹ️ Context

This has been failing since the AWS v2 migration a few weeks back.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

Local linting and testing. Will need to be actually deployed to be fully
vetted.