BCDA-9671: Update deploy-all worker version check (#1286)
## 🎫 Ticket

https://jira.cms.gov/browse/BCDA-9671

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->

- re-enable the worker version check in the deploy-all workflow
- update the worker version check to check the version of the tag used
by the worker service

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

There are checks for each service that ensure the correct version of the
service has been deployed. The api and ssas services have endpoints that
provide this info, but the worker version must be fetched another way.
This PR intends to re-enable this version check and update it to grab
the version from the image associated with the service's running task.

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

Tested with [deployment to
dev](https://github.com/CMSgov/bcda-app/actions/runs/20601880004)